+++ conf/views/whatIsProtectorDateOfBirth.njk
@@ -27,7 +27,7 @@
           }) }}
         {% endif %}
 
-        <form action="{{ routes.controllers.ProtectorDateOfBirthController.onSubmit(mode).url }}" method="post">
+        <form action="{{ routes.controllers.WhatIsProtectorDateOfBirthController.onSubmit(mode).url }}" method="post">
 
           {{ csrf() | safe }}
