+++ test/pages/WhatIsProtectorDateOfBirthPageSpec.scala
@@ -21,18 +21,18 @@ import java.time.LocalDate
 import org.scalacheck.Arbitrary
 import pages.behaviours.PageBehaviours
 
-class ProtectorDateOfBirthPageSpec extends PageBehaviours {
+class WhatIsProtectorDateOfBirthPageSpec extends PageBehaviours {
 
-  "ProtectorDateOfBirthPage" - {
+  "What Is Protector Date Of Birth Page" - {
 
     implicit lazy val arbitraryLocalDate: Arbitrary[LocalDate] = Arbitrary {
       datesBetween(LocalDate.of(1900, 1, 1), LocalDate.of(2100, 1, 1))
     }
 
-    beRetrievable[LocalDate](ProtectorDateOfBirthPage)
+    beRetrievable[LocalDate](WhatIsProtectorDateOfBirthPage)
 
-    beSettable[LocalDate](ProtectorDateOfBirthPage)
+    beSettable[LocalDate](WhatIsProtectorDateOfBirthPage)
 
-    beRemovable[LocalDate](ProtectorDateOfBirthPage)
+    beRemovable[LocalDate](WhatIsProtectorDateOfBirthPage)
   }
 }
